# Update sensor code with speed and filespace improvements.