@erichiggins 2019-02-15
@jveitchmichaelis Have any feedback on this before I merge it?