@erichiggins 2019-02-24
@jveitchmichaelis FYI, I posted an [article to share the benefits of this update](https://blog.ammolytics.com/2019-02-23/rifle-accelerometer-update.html) and gave you a shout out. Hope you don't mind.